Baking the SoA classes into the mesh class
In order for an efficient API, we need both the problem data, and
element data struct of arrays to be accessible in one struct. This
will imply that we will set up a second class hierarchy mimicing the
current element mesh hierarchy. From here, vectorized functions will
be called directly on the SoAs.